backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-generativeai
python-multipart
pypdf2
pydantic
python-dotenv
numpy
